refactor(pooled): extract shared key helpers, move numpy engine behind an engine switch
